Organize project now that we deal with two encodings..
