+++ lib/bech32.js
@@ -1,11 +1,7 @@
 'use strict';
 
-var BigInteger = require('bigi');
 var ALPHABET = 'qpzry9x8gf2tvdw0s3jn54khce6mua7l';
-
 // pre-compute lookup table
-var SEPARATOR = ':';
-var CSLEN = 8;
 var ALPHABET_MAP = {};
 for (var z = 0; z < ALPHABET.length; z++) {
   var x = ALPHABET.charAt(z);
@@ -15,42 +11,23 @@ for (var z = 0; z < ALPHABET.length; z++) {
 }
 
 function polymodStep(pre) {
-  // 29 bits left
-  var b = pre.shiftRight(35);
-  var mask = BigInteger.fromHex('07ffffffff');
-
-  var v = pre.and(mask).shiftLeft(new BigInteger('5'));
-
-  if (b.and(new BigInteger('1')).intValue() > 0) {
-    v = v.xor(BigInteger.fromHex('98f2bc8e61'));
-  }
-  if (b.and(new BigInteger('2')).intValue()) {
-    v = v.xor(BigInteger.fromHex('79b76d99e2'));
-  }
-  if (b.and(new BigInteger('4')).intValue()) {
-    v = v.xor(BigInteger.fromHex('f33e5fb3c4'));
-  }
-  if (b.and(new BigInteger('8')).intValue()) {
-    v = v.xor(BigInteger.fromHex('ae2eabe2a8'));
-  }
-  if (b.and(new BigInteger('16')).intValue()) {
-    v = v.xor(BigInteger.fromHex('1e4f43e470'));
-  }
-
-  return v;
+  var b = pre >> 25;
+  return (pre & 0x1FFFFFF) << 5 ^ -(b >> 0 & 1) & 0x3b6a57b2 ^ -(b >> 1 & 1) & 0x26508e6d ^ -(b >> 2 & 1) & 0x1ea119fa ^ -(b >> 3 & 1) & 0x3d4233dd ^ -(b >> 4 & 1) & 0x2a1462b3;
 }
-
 function prefixChk(prefix) {
-  var chk = new BigInteger('1');
+  var chk = 1;
   for (var i = 0; i < prefix.length; ++i) {
     var c = prefix.charCodeAt(i);
+    if (c < 33 || c > 126) throw new Error('Invalid prefix (' + prefix + ')');
 
-    var mixwith = new BigInteger('' + (c & 0x1f));
-    chk = polymodStep(chk).xor(mixwith);
+    chk = polymodStep(chk) ^ c >> 5;
   }
-
   chk = polymodStep(chk);
 
+  for (var _i = 0; _i < prefix.length; ++_i) {
+    var v = prefix.charCodeAt(_i);
+    chk = polymodStep(chk) ^ v & 0x1f;
+  }
   return chk;
 }
 
@@ -61,24 +38,23 @@ function encode(prefix, words) {
 
   // determine chk mod
   var chk = prefixChk(prefix);
-  var result = prefix + SEPARATOR;
+  var result = prefix + '1';
   for (var i = 0; i < words.length; ++i) {
     var _x = words[i];
-    if (_x >>> 5 !== 0) throw new Error('Non 5-bit word');
+    if (_x >> 5 !== 0) throw new Error('Non 5-bit word');
 
-    chk = polymodStep(chk).xor(new BigInteger('' + _x));
+    chk = polymodStep(chk) ^ _x;
     result += ALPHABET.charAt(_x);
   }
 
-  for (var _i = 0; _i < CSLEN; ++_i) {
+  for (var _i2 = 0; _i2 < 6; ++_i2) {
     chk = polymodStep(chk);
   }
-  chk = chk.xor(new BigInteger('1'));
+  chk ^= 1;
 
-  for (var _i2 = 0; _i2 < CSLEN; ++_i2) {
-    var pos = 5 * (CSLEN - 1 - _i2);
-    var v2 = chk.shiftRight(new BigInteger('' + pos)).and(BigInteger.fromHex('1f'));
-    result += ALPHABET.charAt(v2.toString(10));
+  for (var _i3 = 0; _i3 < 6; ++_i3) {
+    var v = chk >> (5 - _i3) * 5 & 0x1f;
+    result += ALPHABET.charAt(v);
   }
 
   return result;
@@ -94,8 +70,7 @@ function decode(str) {
   if (str !== lowered && str !== uppered) throw new Error('Mixed-case string ' + str);
   str = lowered;
 
-  var split = str.lastIndexOf(SEPARATOR);
-  if (split === -1) throw new Error('No separator character for ' + str);
+  var split = str.lastIndexOf('1');
   if (split === 0) throw new Error('Missing prefix for ' + str);
 
   var prefix = str.slice(0, split);
@@ -108,13 +83,14 @@ function decode(str) {
     var c = wordChars.charAt(i);
     var v = ALPHABET_MAP[c];
     if (v === undefined) throw new Error('Unknown character ' + c);
-    chk = polymodStep(chk).xor(new BigInteger('' + v));
+    chk = polymodStep(chk) ^ v;
+
     // not in the checksum?
-    if (i + CSLEN >= wordChars.length) continue;
+    if (i + 6 >= wordChars.length) continue;
     words.push(v);
   }
 
-  if (chk.toString(10) !== '1') throw new Error('Invalid checksum for ' + str);
+  if (chk !== 1) throw new Error('Invalid checksum for ' + str);
   return { prefix: prefix, words: words };
 }
 
@@ -130,7 +106,7 @@ function convert(data, inBits, outBits, pad) {
 
     while (bits >= outBits) {
       bits -= outBits;
-      result.push(value >>> bits & maxV);
+      result.push(value >> bits & maxV);
     }
   }
