+++ lib/index.js
@@ -1,133 +1,5 @@
 'use strict';
 
-var ALPHABET = 'qpzry9x8gf2tvdw0s3jn54khce6mua7l';
-// pre-compute lookup table
-var ALPHABET_MAP = {};
-for (var z = 0; z < ALPHABET.length; z++) {
-  var x = ALPHABET.charAt(z);
-
-  if (ALPHABET_MAP[x] !== undefined) throw new TypeError(x + ' is ambiguous');
-  ALPHABET_MAP[x] = z;
-}
-
-function polymodStep(pre) {
-  var b = pre >> 25;
-  return (pre & 0x1FFFFFF) << 5 ^ -(b >> 0 & 1) & 0x3b6a57b2 ^ -(b >> 1 & 1) & 0x26508e6d ^ -(b >> 2 & 1) & 0x1ea119fa ^ -(b >> 3 & 1) & 0x3d4233dd ^ -(b >> 4 & 1) & 0x2a1462b3;
-}
-function prefixChk(prefix) {
-  var chk = 1;
-  for (var i = 0; i < prefix.length; ++i) {
-    var c = prefix.charCodeAt(i);
-    if (c < 33 || c > 126) throw new Error('Invalid prefix (' + prefix + ')');
-
-    chk = polymodStep(chk) ^ c >> 5;
-  }
-  chk = polymodStep(chk);
-
-  for (var _i = 0; _i < prefix.length; ++_i) {
-    var v = prefix.charCodeAt(_i);
-    chk = polymodStep(chk) ^ v & 0x1f;
-  }
-  return chk;
-}
-
-function encode(prefix, words) {
-  // too long?
-  if (prefix.length + 7 + words.length > 90) throw new TypeError('Exceeds Bech32 maximum length');
-  prefix = prefix.toLowerCase();
-
-  // determine chk mod
-  var chk = prefixChk(prefix);
-  var result = prefix + '1';
-  for (var i = 0; i < words.length; ++i) {
-    var _x = words[i];
-    if (_x >> 5 !== 0) throw new Error('Non 5-bit word');
-
-    chk = polymodStep(chk) ^ _x;
-    result += ALPHABET.charAt(_x);
-  }
-
-  for (var _i2 = 0; _i2 < 6; ++_i2) {
-    chk = polymodStep(chk);
-  }
-  chk ^= 1;
-
-  for (var _i3 = 0; _i3 < 6; ++_i3) {
-    var v = chk >> (5 - _i3) * 5 & 0x1f;
-    result += ALPHABET.charAt(v);
-  }
-
-  return result;
-}
-
-function decode(str) {
-  if (str.length < 8) throw new TypeError(str + ' too short');
-  if (str.length > 90) throw new TypeError(str + ' too long');
-
-  // don't allow mixed case
-  var lowered = str.toLowerCase();
-  var uppered = str.toUpperCase();
-  if (str !== lowered && str !== uppered) throw new Error('Mixed-case string ' + str);
-  str = lowered;
-
-  var split = str.lastIndexOf('1');
-  if (split === 0) throw new Error('Missing prefix for ' + str);
-
-  var prefix = str.slice(0, split);
-  var wordChars = str.slice(split + 1);
-  if (wordChars.length < 6) throw new Error('Data too short');
-
-  var chk = prefixChk(prefix);
-  var words = [];
-  for (var i = 0; i < wordChars.length; ++i) {
-    var c = wordChars.charAt(i);
-    var v = ALPHABET_MAP[c];
-    if (v === undefined) throw new Error('Unknown character ' + c);
-    chk = polymodStep(chk) ^ v;
-
-    // not in the checksum?
-    if (i + 6 >= wordChars.length) continue;
-    words.push(v);
-  }
-
-  if (chk !== 1) throw new Error('Invalid checksum for ' + str);
-  return { prefix: prefix, words: words };
-}
-
-function convert(data, inBits, outBits, pad) {
-  var value = 0;
-  var bits = 0;
-  var maxV = (1 << outBits) - 1;
-
-  var result = [];
-  for (var i = 0; i < data.length; ++i) {
-    value = value << inBits | data[i];
-    bits += inBits;
-
-    while (bits >= outBits) {
-      bits -= outBits;
-      result.push(value >> bits & maxV);
-    }
-  }
-
-  if (pad) {
-    if (bits > 0) {
-      result.push(value << outBits - bits & maxV);
-    }
-  } else {
-    if (bits >= inBits) throw new Error('Excess padding');
-    if (value << outBits - bits & maxV) throw new Error('Non-zero padding');
-  }
-
-  return result;
-}
-
-function toWords(bytes) {
-  return convert(bytes, 8, 5, true);
-}
-
-function fromWords(words) {
-  return convert(words, 5, 8, false);
-}
-
-module.exports = { decode: decode, encode: encode, toWords: toWords, fromWords: fromWords };
+var bech32 = require('./bech32');
+var base32 = require('./base32');
+module.exports = { bech32: bech32, base32: base32 };
