+++ test/fixtures/base32.json
@@ -20,6 +20,18 @@
         "words": [ 0, 1, 27, 10, 0, 16, 2, 19, 23, 22, 16, 10, 17, 2, 30, 26, 10, 5, 27, 27, 16, 26, 16, 21, 24, 14, 25, 9, 30, 21, 12, 24, 14, 12 ]
       },
       {
+        "string": "bchtest:ppm2qsznhks23z7629mms6s4cwef74vcwvhanqgjxu",
+        "prefix": "bchtest",
+        "hex": "0876a04053bda0a88bda5177b86a15c3b29f559873",
+        "words": [ 1, 1, 27, 10, 0, 16, 2, 19, 23, 22, 16, 10, 17, 2, 30, 26, 10, 5, 27, 27, 16, 26, 16, 21, 24, 14, 25, 9, 30, 21, 12, 24, 14, 12 ]
+      },
+      {
+        "string": "bchreg:pqzg22ty3m437frzk4y0gvvyqj02jpfv7udqugqkne",
+        "prefix": "bchreg",
+        "hex": "08048529648eeb1f2462b548f43184049ea9052cf7",
+        "words": [ 1, 0, 2, 8, 10, 10, 11, 4, 17, 27, 21, 17, 30, 9, 3, 2, 22, 21, 4, 15, 8, 12, 12, 4, 0, 18, 15, 10, 18, 1, 9, 12, 30, 28 ]
+      },
+      {
         "string": "bitcoincash:qr95sy3j9xwd2ap32xkykttr4cvcu7as4y0qverfuy",
         "prefix": "bitcoincash",
         "hex": "00cb481232299cd5743151ac4b2d63ae198e7bb0a9",
@@ -51,7 +63,14 @@
       }
     ],
     "invalid": [
-
+      {
+        "string": "bitcoincashqpm2qsznhks23z7629mms6s4cwef74vcwvy22gdx6a",
+        "exception": "No separator character for bitcoincashqpm2qsznhks23z7629mms6s4cwef74vcwvy22gdx6a"
+      },
+      {
+        "string": "bitcoincash:qpm2qsznhks23z7629mms6s4cwef74vcwvy22gdx6z",
+        "exception": "Invalid checksum for bitcoincash:qpm2qsznhks23z7629mms6s4cwef74vcwvy22gdx6z"
+      }
     ]
   }
 }
